% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/challenges.R
\name{deux_cpp}
\alias{deux_cpp}
\title{Second}
\usage{
deux_cpp(x)
}
\arguments{
\item{vec}{Numerical vector with no missing values}
}
\value{
A numerical vector containing the longest continuous increasing subset
}
\description{
Given a vector gives the longest continuous increasing subset
}
