% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/challenges.R
\name{trois}
\alias{trois}
\title{Third}
\usage{
trois(vec)
}
\arguments{
\item{vec}{Numerical vector}
}
\value{
A single numerical vector with counts of each unique element
}
\description{
Given a vector return the count of each unique element

Hint: Try looking into \code{tabulate}, \code{fastmatch::fastmatch}
}
