% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/challenges.R
\name{un}
\alias{un}
\title{First function}
\usage{
un(m)
}
\arguments{
\item{m}{a square matrix with no missing values}
}
\value{
Single numerical value
}
\description{
Given a square matrix, calculates the average over the sum of row averages and column averages
}
\examples{
m <- matrix(seq(16),nrow=4)
un(m)
}
